database seeded, drop downs functionally on layout enum

+++ Async_Inn/Async_Inn/Models/HotelRoom.cs
@@ -9,7 +9,7 @@ namespace Async_Inn.Models
     {
         public int HotelID { get; set; }
         public int RoomNumber { get; set; }
-        public decimal RoomID { get; set; }
+        public int RoomID { get; set; }
         public decimal Rate { get; set; }
         public bool PetFriendly { get; set; }
